Refactor ClientBaseCase test + update build file for TLS mode integration tests
