+++ src/it/java/net/spy/memcached/util/ConnectionUtilLinuxImpl.java
@@ -31,14 +31,8 @@ public class ConnectionUtilLinuxImpl implements ConnectionUtil {
 	 */
 	@Override
 	public void addLocalMemcachedServer(int...ports) throws Exception {
-		if (TestConfig.isTlsMode()){
-			for(int p: ports) {
-				addLocalMemcachedServer(TestConfig.MEMCACHED_PATH, p, getMemcachedTlsParams());		  
-			}
-		} else {
-			for(int p: ports) {
-				addLocalMemcachedServer(TestConfig.MEMCACHED_PATH, p, getMemcachedParams());	
-			}
+		for(int p: ports) {
+			addLocalMemcachedServer(TestConfig.MEMCACHED_PATH, p, getMemcachedParams());
 		}
 	}
 
@@ -146,6 +140,8 @@ public class ConnectionUtilLinuxImpl implements ConnectionUtil {
 	 * Params:
 	 * username = memcached_user
 	 * memory limit = 4MB 
+	 * enable dynamic reports for 'stats sizes' command
+	 * set certificate and key if TLS/SSL enabled
 	 * 
 	 * @return
 	 */
@@ -153,28 +149,17 @@ public class ConnectionUtilLinuxImpl implements ConnectionUtil {
 	@Override
 	public Map<String,String> getMemcachedParams() {
 		Map<String,String> params = new HashMap<String,String>();
-		params.put("-m", "4");
-		params.put("-o", "track_sizes");
-		return params;
-	}
 
-	/**
-	 * Params:
-	 * username = memcached_user
-	 * memory limit = 4MB
-	 * enable dynamic reports for 'stats sizes' command
-	 * enable TLS/SSL with certificate and key
-	 * 
-	 * @return
-	 */
-	@Override
-	public Map<String,String> getMemcachedTlsParams() {
-		Map<String,String> params = new HashMap<String,String>();
-		String publicCert = "ssl_chain_cert=" + TestConfig.MEMCACHED_CERT + "/public.cert";
-		String privateCert = "ssl_key=" + TestConfig.MEMCACHED_CERT + "/private.cert";
-		params.put("-Z -o", publicCert + "," + privateCert);
-		params.put("-m", "4");
-		params.put("-o", "track_sizes");
+		if (TestConfig.isTlsMode()){
+			String publicCert = "ssl_chain_cert=" + TestConfig.MEMCACHED_CERT_FOLDER + "/public.cert";
+			String privateCert = "ssl_key=" + TestConfig.MEMCACHED_CERT_FOLDER + "/private.cert";
+			params.put("-Z -o", publicCert + "," + privateCert);
+			params.put("-m", "4");
+			params.put("-o", "track_sizes");
+		} else {
+			params.put("-m", "4");
+			params.put("-o", "track_sizes");
+		}
 		return params;
 	}
